==> reports/2_CleanOTU.Rmd <==
---
title: "2_Clean OTUs"
author: "Sylvain Piry, Nicolas Rode"
date: "`r format(Sys.Date(), '%d-%B-%Y')`"
output: 
  html_document:
    theme: "journal"
    toc: true
    toc_depth: 3
    number_sections: true
    toc_float:
      collapsed: false
      smooth_scroll: false
editor_options: 
  chunk_output_type: console
---


```{r setup, include =FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
devtools::load_all()
```

# Clean 16S
```{r }
## Follow the instructions to install dada2: https://benjjneb.github.io/dada2/

clean_OTU(abundance_table="16S_Galaxy17-[FROGS_BIOM_to_TSV__abundance.tsv].tsv", affiliation_table="16S_Galaxy18-[FROGS_BIOM_to_TSV__multi-affiliations.tsv].tsv", thresh=0.5, name = "2022_16S_JRLAT", removecol="NC-Index")


```
# Clean COI
```{r }
## Import abundance table
abundance_table <- data.table::fread(file=file.path(here::here(), "data", "raw_data", "COI_Galaxy22-[FROGS_BIOM_to_TSV__abundance.tsv].tsv"),sep="\t",header=TRUE)# to modify (name of the abundance file)

## Import corrected affiliation table
corrected_affiliation_table <- data.table::fread(file=file.path(here::here(), "data", "raw_data", "COI_Galaxy23-FROGS_BIOM_to_TSV__monoaffiliations.csv"),sep=";",header=TRUE)# to modify (name of the abundance file)
head(corrected_affiliation_table)

## Find cluster with multiple affiliation
multi <- abundance_table$observation_name[grep("Multi", abundance_table$blast_taxonomy)]

## Correct affiliation
abundance_table$blast_taxonomy[grep("Multi", abundance_table$blast_taxonomy)] <- corrected_affiliation_table$blast_taxonomy[match(multi, corrected_affiliation_table$`#observation_name`)]

find_chimeras(fordada=abundance_table, name = "2022_COI_JRLAT")

```
