==> reports/3_FilterCOI.Rmd <==
---
title: "3_FilterCOI"
author: "Nicolas Rode"
date: "`r format(Sys.Date(), '%d-%B-%Y')`"
output: 
  html_document:
    theme: "journal"
    toc: true
    toc_depth: 3
    number_sections: true
    toc_float:
      collapsed: false
      smooth_scroll: false
editor_options: 
  chunk_output_type: console
---


```{r setup, include =FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
devtools::load_all()
```

# Check coverage of different OTUs/samples
```{r }
## Import clean data
abundance_table <- data.table::fread(file=file.path(here::here(), "data", "derived_data", "2022_COI_JRLAT_cleaned_abundance.tsv"),sep="\t",header=TRUE)

## Split taxonomy and add species as row name
abundance_table2 <- abundance_table %>% 
  tidyr::separate(col=blast_taxonomy, into=c("Domain","Phylum", "Class", "Order", "Family", "Genus", "Species"), sep=";", remove=FALSE) %>%
  dplyr::mutate(Species=paste0(Species, rownames(.)), Species2=Species) %>%
  tibble::column_to_rownames(var = "Species2")
## Check rownames
rownames(abundance_table2)

## Check distribution of coverage depth across OTUs
ggplot(abundance_table2, aes(x=observation_sum)) + geom_histogram()+ scale_x_log10()

## Check OTUs with the highest coverage
abundance_table2 %>% 
  dplyr::select(blast_taxonomy, observation_sum) %>% 
  dplyr::filter(observation_sum > 100000)

## Check OTUs with the lowest coverage
abundance_table2 %>% 
  dplyr::select(observation_sum) %>% 
  dplyr::filter(observation_sum < 100000)

## Check samples with the highest coverage
covdepthbysample <- abundance_table2 %>% 
  dplyr::summarise(across(contains("-A"), sum))
## Transpose the dataset
covdepthbysample <- dplyr::as_tibble(data.frame(barcode = names(covdepthbysample), covdepth=t(covdepthbysample)))

## Check distribution of coverage depth across samples
ggplot(covdepthbysample, aes(x=covdepth+1)) + geom_histogram() + scale_x_log10() + geom_vline(aes(xintercept = 1000), colour="red")

## Check samples with the highest coverage=pupae, larvae and adults
covdepthbysample %>% 
  dplyr::filter(covdepth > 1000) %>%
  print(n = 100)

## Check samples with the lowest coverage=pupal case, larvae and fruits
covdepthbysample %>% 
  dplyr::filter(covdepth < 1000)  %>% print(n = 100)

grepl("Species",x)
```

# Filter abundances based on negative controls
```{r }

## Use controls to find maximum coverage depth for each OTU
## See section "Threshold TCC: filtering for cross-contamination" p9 in Galan et al 2016: https://journals.asm.org/doi/10.1128/mSystems.00032-16
## Find controls
CT <- abundance_table2 %>% 
  dplyr::select(Species, contains("CT")&contains("-A")&(contains("PCR")|contains("EXT")))

## Transform in long format to find the maximum coverage depth of each OTUs across all controls
## Sort final dataset so that OTUs are in the right order
CT <- CT %>% 
 tidyr::pivot_longer(!Species,
    names_to = c("Barcode")) %>%
  dplyr::group_by(Species) %>%
  dplyr::slice(which.max(value)) %>%
  dplyr::arrange(factor(Species, levels = abundance_table2$Species))  %>%
  print(n = 30)


Filter_contaminant <- function(x, TccTfa){
  x[x<=TccTfa] <- 0
  return(x)
}
## Filter abundance table
abundance_tableTccfiltered <- abundance_table %>%
  dplyr::mutate(dplyr::across(ends_with("-A")|ends_with("-B")|ends_with("-C"), ~Filter_contaminant(., CT$value)))

```

# Filter abundances based on number of sequences on the flow cell
```{r }
## See section "Threshold TFA: filtering out incorrectly assigned sequences" p9 in Galan et al 2016: https://journals.asm.org/doi/10.1128/mSystems.00032-16
## We used a value of Rfa of 0.01%, which is a bit less conservative than the 0.02% of Galan et al 2016
Rfa <- 0.0001

## Filter abundance table
abundance_tableTccTfafiltered <- abundance_tableTccfiltered %>%
  dplyr::mutate(dplyr::across(ends_with("-A")|ends_with("-B")|ends_with("-C"), ~Filter_contaminant(., abundance_tableTccfiltered$observation_sum*Rfa)))

## Export results
write.table(abundance_tableTccTfafiltered, file=file.path(here::here(), "data", "derived_data", "2022_COI_JRLAT_cleaned_filtered_abundance_final.tsv"), sep="\t", row.names=FALSE, col.names=TRUE, quote=FALSE)
```

# Update metadata
```{r, eval=TRUE}
## Import metadata table
metadata <- data.table::fread(file=file.path(here::here(), "data", "raw_data", "Sample_ISS.xlsx - Metadata.csv"), sep=",", header=TRUE)
tail(metadata)

##Remove the two last empty rows
metadata <- metadata %>%
  dplyr::filter(Date != "Not done")
tail(metadata)

## Sum coverage depth across OTUs from the same species
abundance_table_sum <- abundance_tableTccTfafiltered %>% dplyr::select(blast_taxonomy, observation_sum:Undetermined) %>% 
  dplyr::group_by(blast_taxonomy) %>% 
  dplyr::summarise(across(everything(), sum))

##Compute proportion only for rows with Drosophila and columns from replicate A
prop_table_clean <- abundance_table_sum %>%
  dplyr::filter(grepl("Drosophila", blast_taxonomy)) %>%
  dplyr::select(blast_taxonomy:Undetermined) %>% 
  dplyr::select(blast_taxonomy, ends_with("-A")) %>%  dplyr::mutate_if(is.numeric, ~round(./sum(.), 1)) 

## Transpose the table
prop_table_clean <- prop_table_clean %>%
    tidyr::gather(key = Barcode, value = value, 2:ncol(prop_table_clean)) %>% 
    tidyr::spread(key = names(prop_table_clean)[1],value = 'value')

## Simplify the name
colnames(prop_table_clean) <- gsub("Arthropoda;Insecta;Diptera;Drosophilidae;Drosophilinae;Drosophila;", "", colnames(prop_table_clean))
## Put back the barcode name
prop_table_clean$Barcode <- gsub("-", "_", prop_table_clean$Barcode)
prop_table_clean$Barcode <- gsub("_A", "", prop_table_clean$Barcode)

## Remove the Drosophila yakuba whose reads are at a very low frequency in the samples
prop_table_clean$Drosophila_yakuba
prop_table_clean <- prop_table_clean %>%
  dplyr::select(-Drosophila_yakuba)

## Merge the two datasets
metadata2 <-  dplyr::left_join(metadata, prop_table_clean, by="Barcode")

write.table(metadata2, file=file.path(here::here(), "data", "derived_data", "Metadata_final.tsv"), sep="\t", row.names=FALSE, col.names=TRUE, quote=FALSE)
```
