==> reports/4_Filter16S.Rmd <==
---
title: "4_Filter16S"
author: "Nicolas Rode"
date: "`r format(Sys.Date(), '%d-%B-%Y')`"
output: 
  html_document:
    theme: "journal"
    toc: true
    toc_depth: 3
    number_sections: true
    toc_float:
      collapsed: false
      smooth_scroll: false
editor_options: 
  chunk_output_type: console
---


```{r setup, include =FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
devtools::load_all()
```

# Check coverage of different OTUs/samples
```{r }
## Import clean data
abundance_table <- data.table::fread(file=file.path(here::here(), "data", "derived_data", "2022JRLAT-16S_cleaned_abundance.txt"),sep="\t",header=TRUE)

2## Remove chloroplast and mitochondria
abundance_table <- abundance_table %>%
  dplyr::filter(!grepl("Plantae", blast_taxonomy))

## Split taxonomy and add species as row name
abundance_table2 <- abundance_table %>% 
  tidyr::separate(col=blast_taxonomy, into=c("Domain","Phylum", "Class", "Order", "Family", "Genus", "Species"), sep=";", remove=FALSE) %>%
  dplyr::mutate(Genus=paste0(Genus, rownames(.)), Genus2=Genus) %>%
  tibble::column_to_rownames(var = "Genus2")
## Check rownames
rownames(abundance_table2)

## Check distribution of coverage depth across OTUs
ggplot(abundance_table2, aes(x=observation_sum)) + geom_histogram() + scale_x_log10()

## Check OTUs with the highest coverage
abundance_table2 %>% 
  dplyr::select(blast_taxonomy, Species, observation_sum) %>% 
  dplyr::filter(observation_sum > 100000)

## Check OTUs with the lowest coverage
abundance_table2 %>% 
  dplyr::select(observation_sum) %>% 
  dplyr::filter(observation_sum < 100000)

## Check samples with the highest coverage
covdepthbysample <- abundance_table2 %>% 
  dplyr::summarise(across(contains("-A")|contains("-B")|contains("-C"), sum))

## Transpose the dataset
covdepthbysample <- dplyr::as_tibble(data.frame(barcode = names(covdepthbysample), covdepth=t(covdepthbysample)))

## Check distribution of coverage depth across samples
ggplot(covdepthbysample, aes(x=covdepth+1)) + geom_histogram() + geom_vline(aes(xintercept = 7000), colour="red")

## Check samples with the highest coverage
covdepthbysample %>% 
  dplyr::filter(covdepth > 7000) %>%
  print(n = 100)

## Check samples with the lowest coverage=pupal case, larvae and fruits
covdepthbysample %>% 
  dplyr::filter(covdepth < 7000)  %>%
  print(n = 115)

```

# Filter abundances based on negative controls
```{r }

## Use controls to find maximum coverage depth for each OTU
## See section "Threshold TCC: filtering for cross-contamination" p9 in Galan et al 2016: https://journals.asm.org/doi/10.1128/mSystems.00032-16
## Find controls
CT <- abundance_table2 %>% 
  dplyr::select(Genus, contains("CT")&(contains("PCR")|contains("EXT")))

## Transform in long format to find the maximum coverage depth of each OTUs across all controls
## Sort final dataset so that OTUs are in the right order
CT <- CT %>% 
 tidyr::pivot_longer(!Genus,
    names_to = c("Barcode")) %>%
  dplyr::group_by(Genus) %>%
  dplyr::slice(which.max(value)) %>%
  dplyr::arrange(factor(Genus, levels = abundance_table2$Genus))  %>%
  print(n = 30)


Filter_contaminant <- function(x, Tcc_Tfa){
  x[x<=Tcc_Tfa] <- 0
  return(x)
}
## Filter abundance table
abundance_tableTccfiltered <- abundance_table %>%
  dplyr::mutate(dplyr::across(ends_with("-A")|ends_with("-B")|ends_with("-C"), ~Filter_contaminant(., CT$value)))

```
# Filter abundances based on number of sequences on the flow cell
```{r }
## See section "Threshold TFA: filtering out incorrectly assigned sequences" p9 in Galan et al 2016: https://journals.asm.org/doi/10.1128/mSystems.00032-16
## We used a value of Rfa of 0.01%, which is a bit less conservative than the 0.02% of Galan et al 2016
Rfa <- 0.0001

## Filter abundance table
abundance_tableTccTfafiltered <- abundance_tableTccfiltered %>%
  dplyr::mutate(dplyr::across(ends_with("-A")|ends_with("-B")|ends_with("-C"), ~Filter_contaminant(., abundance_tableTccfiltered$observation_sum*Rfa)))

## Export results
write.table(abundance_tableTccTfafiltered, file=file.path(here::here(), "data", "derived_data", "2022_16S_JRLAT_cleaned_filtered_abundance_final.tsv"), sep="\t", row.names=FALSE, col.names=TRUE, quote=FALSE)

```
# Compare observed and observed frequency in the mock community (DNA standard)
```{r }

## Find Mock communities (only replicates A and C, since B did not work)
CT1 <- abundance_table2 %>% 
  dplyr::select(Order, Genus, contains("CTL-DN-MOC-COM-01-A")|contains("CTL-DN-MOC-COM-01-C"))

## Transform in long format to find the maximum coverage depth of each OTUs across each mock
## Sort final dataset so that OTUs are in the right order
CT1 <- CT1 %>% 
 tidyr::pivot_longer(!c(Order, Genus),
    names_to = c("Barcode")) %>%
  dplyr::group_by(Order, Genus) %>%
  dplyr::summarise(Mean=mean(value)) %>%
  dplyr::arrange(dplyr::desc(Mean)) %>%
  print(n = 30)

## Compute observed frequencies
CT1 <- CT1 %>%
  dplyr::ungroup() %>%
  dplyr::filter(dplyr::row_number()%in%1:8) %>%
  dplyr::mutate(ObsFreq = round(Mean / sum(Mean), 2))

## Name and expected composition for the 8 bacterial species in the mock:https://www.zymoresearch.com/collections/zymobiomics-microbial-community-standards/products/zymobiomics-microbial-community-dna-standard

theoreticalproportions <- data.frame(Zymobact=c("Pseudomonas aeruginosa","Escherichia coli","Salmonella enterica", "Lactobacillus fermentum", "Enterococcus faecalis", "Staphylococcus aureus", "Listeria monocytogenes", "Bacillus subtilis"), ExpectedComp16S=c(4.2, 10.1,10.4, 18.4, 9.9, 15.5, 14.1, 17.4))

## Change the order
Species <- c("Staphylococcus aureus", "Bacillus subtilis", "Lactobacillus fermentum", "Salmonella enterica","Escherichia coli","Listeria monocytogenes" , "Pseudomonas aeruginosa","Enterococcus faecalis")

theoreticalproportions <- theoreticalproportions[match(Species, theoreticalproportions$Zymobact),]
theoreticalproportions$Prop <- theoreticalproportions$ExpectedComp16S/sum(theoreticalproportions$ExpectedComp16S) 

## Good consistency between observed and expected
MOC_pro=data.frame(CT1, ExpectFreq=theoreticalproportions$Prop, Species=Species )

#test if the manipulation change the proportion 
chisq.test(MOC_pro$ObsFreq,MOC_pro$ExpectFreq)

```
# Compare observed and observed frequency in the mock community (DNA standard II, log distribution)
```{r }
## Find Mock communities (only replicates A and C, since B did not work)
CT2 <- abundance_table2 %>% 
  dplyr::select(Order, Genus, contains("CTL-DN-MOC-COM-02-A")|contains("CTL-DN-MOC-COM-02-C"))

## Transform in long format to find the maximum coverage depth of each OTUs across each mock
## Sort final dataset so that OTUs are in the right order
CT2 <- CT2 %>% 
 tidyr::pivot_longer(!c(Order, Genus),
    names_to = c("Barcode")) %>%
  dplyr::group_by(Order, Genus) %>%
  dplyr::summarise(Mean=mean(value)) %>%
  print(n = 30)

## Compute observed frequencies
CT2 <- CT2 %>%
  dplyr::ungroup() %>%
  dplyr::filter(Genus%in%CT1$Genus) %>%
  dplyr::mutate(ObsFreq = round(Mean / sum(Mean), 2))

## Name and expected composition for the 8 bacterial species in the mock:https://www.zymoresearch.com/collections/zymobiomics-microbial-community-standards/products/zymobiomics-microbial-community-dna-standard

theoreticalproportions <- data.frame(Zymobact=c("Listeria monocytogenes", "Pseudomonas aeruginosa", "Bacillus subtilis","Escherichia coli","Salmonella enterica", "Lactobacillus fermentum", "Enterococcus faecalis", "Staphylococcus aureus"), ExpectedComp16S=c(95.9, 2.8 , 1.2, 0.069, 0.07, 0.012, 0.00067, 0.0001))

## Change the order
Species <- c("Bacillus subtilis","Escherichia coli","Salmonella enterica" ,"Enterococcus faecalis", "Lactobacillus fermentum","Listeria monocytogenes","Staphylococcus aureus", "Pseudomonas aeruginosa")

theoreticalproportions <- theoreticalproportions[match(Species, theoreticalproportions$Zymobact),]
theoreticalproportions$Prop <- round(theoreticalproportions$ExpectedComp16S/sum(theoreticalproportions$ExpectedComp16S), 3)

## Consistency between observed and expected freqeuncy OK
MOC_log=data.frame(CT2, ExpectFreq=theoreticalproportions$Prop, Species=Species )

#test 
chisq.test(MOC_log$ObsFreq,MOC_log$ExpectFreq)
```
