==> reports/6_Analyse_abundance_16S.Rmd <==
---
title: "Analyse_abondance_16S"
author: "GERIN-JEAN David"
date: "2022-12-05"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
devtools::load_all()
```

# Prepare data for phyloseq analysis
## Import data
```{r}
## Please follow the instructions here to prepare your data (without sub-setting samples or taxa): https://vaulot.github.io/tutorials/Phyloseq_tutorial.html#read-the-data-and-create-phyloseq-objects

#table_OTU
#import data
data16S <- data.table::fread(file=file.path(here::here(), "data", "derived_data", "2022_16S_JRLAT_cleaned_filtered_abundance_final.tsv"), sep="\t",header=TRUE)

#import metadata
metadata <- data.table::fread(file=file.path(here::here(), "data", "derived_data", "Metadata_final.tsv"), sep="\t",header=TRUE)

#split to have family names
data16S$blast_taxonomy
taxonomy <- 
  data16S %>%
  dplyr::select(observation_name, blast_taxonomy)  %>%
  tidyr::separate(col=blast_taxonomy, into=c("kingdom","phylum","class","order","family","genus","species"),  sep=";", remove=TRUE)

## Merge unknown families together
table(taxonomy$family)
taxonomy$kingdom[taxonomy$family=="NA"|taxonomy$family=="Multi-affiliation"|taxonomy$family=="unknown family"] <- "Unknown"
taxonomy$phylum[taxonomy$family=="NA"|taxonomy$family=="Multi-affiliation"|taxonomy$family=="unknown family"] <- "Unknown"
taxonomy$class[taxonomy$family=="NA"|taxonomy$family=="Multi-affiliation"|taxonomy$family=="unknown family"] <- "Unknown"
taxonomy$order[taxonomy$family=="NA"|taxonomy$family=="Multi-affiliation"|taxonomy$family=="unknown family"] <- "Unknown"
taxonomy$family[taxonomy$family=="NA"|taxonomy$family=="Multi-affiliation"|taxonomy$family=="unknown family"] <- "Unknown"

```

## Clean metadata
```{r}

## Remove controls and samples from other experiment
metadata <- metadata %>%
  dplyr::filter(!(grepl("^CT", Barcode)|grepl("^ST", Barcode)|grepl("^BL", Barcode)|grepl("^CH", Barcode)))
metadata$Barcode

## Split barcode
metadata <- metadata %>%
  tidyr::separate(col=Barcode, into=c("Type","Fruit","Population","Status","Number"),  sep="_", remove=FALSE)

metadata$Status[metadata$Status=="INM"] <- "INF"
metadata$Status[metadata$Status=="IND"] <- "INF"

metadata$Age <- as.factor(metadata$Status)
levels(metadata$Age) <- c("Fresh", "Fresh", "Old", "NA")
metadata$Status <- as.factor(metadata$Status)
levels(metadata$Status) <- c("Inf", "Uninf", "Uninf", "NA")
  
## Add a column to have more details for the fruits
metadata$Status2 <- as.character(metadata$Status)
metadata$Status2[metadata$Status2=="NA"] <- ""
metadata$Groups <- as.factor(paste(metadata$Type,metadata$Fruit,metadata$Status2))

metadata$Groups <- factor(metadata$Groups, levels= levels(metadata$Groups)[c(2:9, 1)] )
 
  
## mutate(cyl = factor(cyl, levels = c(4, 6, 8)))

colnames(metadata)

## Add a column to know whether the pool is 100% Dsim, 100% Dsuz or a mix of the two species 
metadata$DrosoSpecies <-ifelse(metadata$Type=="FRT"|is.na(metadata$Drosophila_simulans),NA,
ifelse(metadata$Drosophila_simulans>0 & metadata$Drosophila_suzukii==0,
"Dsim", ifelse(metadata$Drosophila_suzukii >0 &
metadata$Drosophila_simulans==0, "Dsuz", "Dim_Dsuz")))

metadata$Species <- ifelse(metadata$Type=="FRT", metadata$Fruit,
paste(metadata$DrosoSpecies, metadata$Fruit, sep="_"))


## Check that the samples for pupal cases and adults are in the same order
metadata$Number[metadata$Type=="PUC"]
metadata$Number[metadata$Type=="ADT"]

## Assign pupal cases the species of the adults that emerged
metadata$Species[metadata$Type=="PUC"] <-
metadata$Species[metadata$Type=="ADT"]
```

## Merge replicates together
```{r}

#select the column start xxx
data16S <- data16S %>%
  dplyr::select(!(dplyr::starts_with("CT")|dplyr::starts_with("ST")|dplyr::starts_with("BL")|dplyr::starts_with("CH")|dplyr::starts_with("Undetermined")|dplyr::starts_with("observation_sum")|dplyr::starts_with("#comment")|dplyr::starts_with("rdp_tax_and_bootstrap")|dplyr::starts_with("seed_id")|dplyr::starts_with("seed_sequence")))


#transform the table to long data set 
data16S <- data16S %>%
  tidyr::pivot_longer(!observation_name, names_to = "Barcode", values_to = "count") 
 

data16S$Samples <- substr(data16S$Barcode, 1, 17)
data16S$Samples <- gsub("-", "_", data16S$Samples)

## Compute mean across the three replicates
data16S <- data16S %>% 
  dplyr::group_by(observation_name, Samples) %>% 
  dplyr::summarise(CountMean = round(mean(count))) %>% 
  dplyr::ungroup()
data16S <- data16S %>% 
  tidyr::pivot_wider(names_from = Samples, values_from = CountMean)

## Remove samples with coverage lower than larvae
## Check samples with the highest coverage
covdepthbysample <- data16S %>% 
  dplyr::select(-observation_name) %>%
  dplyr::summarise_all(sum)
#transform the table to long data set 
covdepthbysample <- covdepthbysample %>%
  tidyr::pivot_longer(cols=everything(), names_to = "Barcode", values_to = "count") %>%
 dplyr::arrange(count) %>%
print(n = 50)

## Check distribution of coverage depth across samples
ggplot(covdepthbysample, aes(x=count)) + geom_histogram() + geom_vline(aes(xintercept = 400), colour="red")+ geom_vline(aes(xintercept = 1300), colour="red")

sampleslowerthan400 <- covdepthbysample$Barcode[covdepthbysample$count<400]
sampleslowerthan1300 <- covdepthbysample$Barcode[covdepthbysample$count<1300]

data16S_higher400 <- data16S %>% 
  dplyr::select(!all_of(sampleslowerthan400))
data16S_higher1300 <- data16S %>% 
  dplyr::select(!all_of(sampleslowerthan1300))

write.table(data16S, file=file.path(here::here(), "data", "derived_data", "2022_16S_JRLAT_cleaned_filtered_abundance_final_perSample.tsv"), sep="\t", row.names=FALSE, col.names=TRUE, quote=FALSE)
```

## Prepare data for phyloseq
```{r}

##Define the row names from the columns
otu_mat <- data16S
otu_mat_higher400 <- data16S_higher400
otu_mat_higher1300 <- data16S_higher1300
colnames(otu_mat)

## Remove OTU column and add it as row names
otu_mat <- otu_mat %>%
    tibble::column_to_rownames(var="observation_name")
otu_mat_higher400 <- otu_mat_higher400 %>%
    tibble::column_to_rownames(var="observation_name")
otu_mat_higher1300 <- otu_mat_higher1300 %>%
    tibble::column_to_rownames(var="observation_name")
rownames(otu_mat_higher400)

## Remove OTU column and add it as row names
tax_mat <- taxonomy %>%
    tibble::column_to_rownames(var="observation_name")
rownames(tax_mat)

## Remove sample column and add it as row names
samples_df <- metadata %>%
    tibble::column_to_rownames(var="Barcode")
rownames(samples_df)

## Transform into matrices otu and tax tables (sample table can be left as data frame)
otu_mat <- as.matrix(otu_mat)
otu_mat_higher400 <- as.matrix(otu_mat_higher400)
otu_mat_higher1300 <- as.matrix(otu_mat_higher1300)
tax_mat <- as.matrix(tax_mat)

## Transform to phyloseq objects
OTU = phyloseq::otu_table(otu_mat, taxa_are_rows = TRUE)
OTU_higher400 = phyloseq::otu_table(otu_mat_higher400, taxa_are_rows = TRUE)
OTU_higher1300 = phyloseq::otu_table(otu_mat_higher1300, taxa_are_rows = TRUE)
TAX = phyloseq::tax_table(tax_mat)
samples = phyloseq::sample_data(samples_df)

phyloseq::sample_names(OTU_higher400)
phyloseq::sample_names(samples)

carbom <- phyloseq::phyloseq(OTU, TAX, samples)
carbom_higher400 <- phyloseq::phyloseq(OTU_higher400, TAX, samples)
carbom_higher1300 <- phyloseq::phyloseq(OTU_higher1300, TAX, samples)
carbom
```

## Plot abundance bar graph
```{r}
phyloseq::sample_names(carbom)
phyloseq::rank_names(carbom)
phyloseq::sample_variables(carbom)
phyloseq::plot_bar(carbom, fill="family")

#Group by adult, fruit, larva, pupal case and pupa
phyloseq.extended::plot_composition(carbom, "kingdom", "Bacteria", "family", fill = "family", numberOfTaxa = 11) +
  facet_grid(~Type, scales = "free_x", space = "free_x")

phyloseq.extended::plot_composition(carbom, "kingdom", "Bacteria", "family", fill = "family", numberOfTaxa = 9) + 
  facet_grid(~Type+Fruit+Status+Population, scales = "free_x", space = "free_x")


p <-phyloseq.extended::plot_composition(carbom_higher400, "kingdom", "Bacteria", "family", fill = "family", numberOfTaxa = 9) + 
  facet_grid(~Type+Fruit, scales = "free_x", space = "free_x") + theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Composition of bacteria family") + 
  labs(fill='Family')

p


ggsave(file.path(here::here(), "plots", "toto.jpeg"), height = 6, width = 8)

#merge sample from the same group together 
#not included in paper 
carbom_type <- phyloseq::merge_samples(carbom, "Groups")
phyloseq::plot_bar(carbom_type, fill = "family") + 
  geom_bar(aes(color = family, fill = family), stat = "identity", position="stack")

## Rarefy data set
carbom_rare <- phyloseq::rarefy_even_depth(carbom, rngseed = 20190124)
carbom_rare_higher400 <- phyloseq::rarefy_even_depth(carbom_higher400, rngseed = 20190124)

## After rarefaction
phyloseq::plot_bar(carbom_rare_higher400, fill = "family")
phyloseq.extended::plot_composition(carbom_rare_higher400, "kingdom", "Bacteria", "family", fill = "family", numberOfTaxa = 9)
phyloseq.extended::plot_composition(carbom_rare_higher400, "kingdom", "Bacteria", "family", fill = "family", numberOfTaxa = 9) + 
  facet_grid(~Type+Fruit+Status+Population, scales = "free_x", space = "free_x") +
  ggtitle("Composition of bacteria family rare_higher400") + 
  labs(fill='Family')


ggsave(file.path(here::here(), "plots", "toto400.jpeg"), height = 6, width = 8)


```
## Rarefaction
```{r}
vegan::rarecurve(t(otu_mat), step=50, cex=0.5)

p <- vegan::rarecurve(t(otu_mat), 
                      step = 50, label = FALSE, 
                      sample = min(rowSums(t(otu_mat)), 
                                   col = "blue", cex = 0.6), ylab ="Number of OTUs")
p

#ggsave(file.path(here::here(), "plots", "Rarefaction.jpeg"), height = 6, width = 8)

```

## Plot alpha diversity by sample
```{r}
## use the code available here: https://vaulot.github.io/tutorials/Phyloseq_tutorial.html#bar-graphs)

phyloseq::plot_richness(carbom_rare, measures=c("Chao1", "Shannon"), color="Groups")
phyloseq::plot_richness(carbom_rare_higher400, measures=c("Chao1", "Shannon"), color="Groups")
```

## Perform ordination (Beta diversity)
```{r, eval=TRUE}
#carbom.ord <- phyloseq::ordinate(carbom, method ="NMDS", distance ="bray")

carbom.ord_carbom_rare_higher400 <- phyloseq::ordinate(carbom_rare_higher400, method ="NMDS", distance ="jaccard")

phyloseq::plot_ordination(carbom_rare_higher400, carbom.ord_carbom_rare_higher400, type="Fruit", color="Groups", shape= "Type", title="Ordination higher than 400") 

```

## Heatmap (Beta diversity)
```{r, eval=TRUE}

total = median(phyloseq::sample_sums(carbom))
carbom_abund_higher400 <- phyloseq::filter_taxa(carbom, function(x) sum(x ) > 400, TRUE)

#check the similarity between the function here and previous splitting
identical(carbom_abund_higher400, carbom_higher400)

phyloseq::plot_heatmap(carbom_rare_higher400, method = "NMDS", distance = "bray", sample.order = "Type", title="Rare higher than 400 method bray")


phyloseq::plot_heatmap(carbom_rare_higher400, method = "NMDS", distance = "jaccard", sample.order = "Groups", title="Rare higher than 400 method jaccard")

```
